tidying up, adding tile borders

+++ Sliding Picture Tiles Game/Surface.h
@@ -41,7 +41,7 @@ public:
 public:
 	Surface( unsigned int width, unsigned int height, unsigned int pitch )
 		:
-		pBuffer( std::make_unique<Color[]>( pitch * height ) ),
+		pBuffer( std::make_unique<Color[]>( (size_t)pitch * height ) ),
 		width( width ),
 		height( height ),
 		pitch( pitch )
@@ -72,13 +72,13 @@ public:
 	{}
 	void Clear( Color fillValue )
 	{
-		memset( pBuffer.get(), fillValue.dword, pitch * height * sizeof( Color ) );
+		memset( pBuffer.get(), fillValue.dword, (size_t)pitch * height * sizeof( Color ) );
 	}
 	void Present( unsigned int dstPitch, BYTE* const pDst ) const
 	{
 		for( unsigned int y = 0; y < height; y++ )
 		{
-			memcpy( &pDst[dstPitch * y], &pBuffer[pitch * y], sizeof( Color ) * width );
+			memcpy( &pDst[dstPitch * y], &pBuffer[(size_t)pitch * y], sizeof( Color ) * width );
 		}
 	}
 	void PutPixel( unsigned int x, unsigned int y, Color c )
@@ -87,7 +87,7 @@ public:
 		assert( y >= 0 );
 		assert( x < width );
 		assert( y < height );
-		pBuffer[y * pitch + x] = c;
+		pBuffer[(size_t)y * pitch + x] = c;
 	}
 	void PutPixelAlpha( unsigned int x, unsigned int y, Color c );
 	Color GetPixel( unsigned int x, unsigned int y ) const
@@ -96,7 +96,7 @@ public:
 		assert( y >= 0 );
 		assert( x < width );
 		assert( y < height );
-		return pBuffer[y * pitch + x];
+		return pBuffer[(size_t)y * pitch + x];
 	}
 	unsigned int GetWidth() const
 	{
